Split character tile into new components
