+++ src/app/character-tile/CharacterTile.scss
@@ -1,0 +1,113 @@
+@import '../variables.scss';
+
+.character-tile {
+  flex: 1;
+  display: flex;
+  position: relative;
+  z-index: 0;
+
+  .background {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    filter: brightness(70%);
+    background-size: cover;
+    background-position: left center;
+    background-repeat: no-repeat;
+    background-color: black;
+    z-index: -1;
+  }
+
+  .character-text {
+    flex: 1;
+    text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5), 0px 0px 10px rgba(0, 0, 0, 0.5);
+    display: flex;
+    flex-direction: column;
+    margin: 4px 6px 2px 0px;
+
+    .vault & {
+      flex-direction: row;
+      align-items: center;
+    }
+
+    .app-icon {
+      filter: drop-shadow(1px 1px 1px rgba(0, 0, 0, 0.5))
+        drop-shadow(0px 0px 10px rgba(0, 0, 0, 0.5));
+    }
+  }
+
+  .emblem {
+    width: 32px;
+    height: 32px;
+    margin: 0 8px;
+    align-self: center;
+    background-color: rgba(0, 0, 0, 0.2);
+    background-size: cover;
+
+    @include phone-portrait {
+      margin-right: 35px;
+    }
+
+    .destiny2 & {
+      visibility: hidden;
+    }
+
+    &.vault {
+      visibility: visible;
+      height: 40px;
+      width: 40px;
+      margin: 8px;
+    }
+  }
+
+  .class,
+  .race-gender {
+    text-transform: capitalize;
+    flex: 1;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    line-height: 1.4em;
+  }
+
+  .powerLevel {
+    font-family: 'Helvetica Neue', 'Helvetica', Arial, sans-serif;
+    font-weight: 500;
+  }
+
+  .powerLevel,
+  .light {
+    color: $power;
+
+    .app-icon {
+      vertical-align: 70%;
+      font-size: 40%;
+      margin-right: 2px;
+    }
+  }
+
+  .level {
+    margin-right: 1px;
+  }
+
+  .top,
+  .bottom {
+    // flex: 1;
+    display: flex;
+    flex-wrap: nowrap;
+    flex-direction: row;
+    align-items: center;
+    font-weight: 400;
+  }
+
+  .top {
+    font-size: 20px;
+    font-weight: 400;
+    height: 24px;
+  }
+
+  .bottom {
+    font-size: 12px;
+    color: white;
+    height: 14px;
+  }
+}
