+++ src/app/destiny1/activities/Activities.tsx
@@ -4,7 +4,7 @@ import BungieImage, { bungieBackgroundStyle } from '../../dim-ui/BungieImage';
 import { DimStore, D1Store } from '../../inventory/store-types';
 import { RootState } from '../../store/reducers';
 import { sortedStoresSelector } from '../../inventory/selectors';
-import SimpleCharacterTile from '../../inventory/SimpleCharacterTile';
+import { CharacterTileButton } from '../../character-tile/CharacterTileButton';
 import CollapsibleTitle from '../../dim-ui/CollapsibleTitle';
 import { AppIcon, starIcon } from '../../shell/icons';
 import { t } from 'app/i18next-t';
@@ -86,7 +86,7 @@ class Activities extends React.Component<Props> {
         <div className="activities-characters">
           {characters.map((store) => (
             <div key={store.id} className="activities-character">
-              <SimpleCharacterTile character={store} />
+              <CharacterTileButton character={store} />
             </div>
           ))}
         </div>
