--- src/app/inventory/StoreHeading.scss
@@ -1,223 +1,0 @@
-@import '../variables.scss';
-
-$xp-bar-height: 2px;
-
-.character {
-  flex: 1;
-  width: 100%;
-}
-
-.powerLevel,
-.light {
-  color: $power;
-}
-
-.character-box {
-  position: relative;
-  border-bottom: $xp-bar-height solid #666;
-  height: 46px;
-  max-width: 230px;
-  cursor: pointer;
-
-  .powerLevel {
-    font-family: 'Helvetica Neue', 'Helvetica', Arial, sans-serif;
-    font-weight: 500;
-  }
-
-  .powerLevel,
-  .light {
-    color: $power;
-
-    .app-icon {
-      vertical-align: 70%;
-      font-size: 40%;
-      margin-right: 2px;
-    }
-  }
-
-  .phone-portrait & {
-    margin: 0 auto;
-  }
-
-  .details {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 4px 6px 4px 8px;
-    position: relative;
-    z-index: 0;
-  }
-
-  .background {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    filter: brightness(70%);
-    background-size: cover;
-    background-position: left center;
-    background-repeat: no-repeat;
-    z-index: 0;
-  }
-
-  .overlay {
-    display: none;
-    z-index: 1;
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    opacity: 0.1;
-    background-color: white;
-  }
-
-  &:hover,
-  &:active {
-    .background {
-      filter: brightness(130%) !important;
-    }
-  }
-
-  .character-text {
-    flex: 1;
-    text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5), 0px 0px 10px rgba(0, 0, 0, 0.5);
-    min-width: 0;
-    display: flex;
-    flex-direction: column;
-    .vault & {
-      flex-direction: row;
-      align-items: center;
-    }
-
-    .app-icon {
-      filter: drop-shadow(1px 1px 1px rgba(0, 0, 0, 0.5))
-        drop-shadow(0px 0px 10px rgba(0, 0, 0, 0.5));
-    }
-  }
-
-  .top,
-  .bottom {
-    width: 100%;
-    display: flex;
-    flex-wrap: nowrap;
-    flex-direction: row;
-    align-items: center;
-    font-weight: 400;
-  }
-  .top {
-    font-size: 20px;
-    font-weight: 400;
-    height: 24px;
-  }
-  .bottom {
-    font-size: 12px;
-    color: white;
-    height: 14px;
-  }
-  .emblem {
-    width: 32px;
-    height: 32px;
-    margin-right: 4px;
-    background-color: rgba(0, 0, 0, 0.2);
-    background-size: cover;
-  }
-  .class,
-  .race-gender {
-    text-transform: capitalize;
-    flex: 1;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    line-height: 1.4em;
-  }
-  .level {
-    margin-right: 1px;
-  }
-
-  &.destiny2 {
-    border-bottom: none;
-
-    .emblem {
-      visibility: hidden;
-    }
-    .background {
-      filter: none;
-    }
-  }
-}
-
-.level-bar {
-  height: $xp-bar-height;
-  width: 100%;
-  position: absolute;
-  bottom: -$xp-bar-height;
-  left: 0;
-  .level-bar-progress {
-    background-color: $xp;
-    height: 100%;
-
-    &.mote-progress {
-      background-color: rgb(40, 132, 179);
-    }
-  }
-}
-
-.character-box.vault {
-  width: calc(6px + var(--character-column-width) - var(--item-margin));
-  border-bottom: 0;
-  .background {
-    background-color: black;
-    background-position: center right;
-    background-size: contain;
-    filter: none;
-  }
-  .emblem {
-    background-color: transparent;
-    width: 40px;
-    height: 40px;
-    margin-right: 6px;
-  }
-  .details {
-    padding-left: 6px;
-    padding-top: 0;
-    padding-bottom: 0;
-    height: 100%;
-  }
-}
-
-.loadout-button {
-  color: white;
-  text-align: center;
-  font-size: 18px !important;
-  cursor: pointer;
-  outline: none;
-  margin-left: 6px;
-  margin-top: 1px;
-  opacity: 0.8;
-}
-
-.current {
-  &.character,
-  &.character-swipe {
-    .character-box:before {
-      content: '';
-      position: absolute;
-      top: 0;
-      left: 0;
-      z-index: 1;
-      border-top: 13px solid gold;
-      border-right: 13px solid transparent;
-    }
-  }
-}
-
-.lang-ru {
-  .vault .class {
-    font-size: 16px;
-  }
-}
-
-.lang-ja {
-  .class {
-    font-size: 16px;
-  }
-}
